# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LeadSprint operator console API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /webhooks/intake:
    post:
      operationId: receiveLeadIntake
      tags: [leads]
      summary: Receive a signed lead intake event
      responses:
        "201": { description: Lead accepted }
        "200": { description: Duplicate lead ignored }
        "401": { description: Invalid signature }
  /webhooks/retell:
    post:
      operationId: receiveRetellEvent
      tags: [calls]
      summary: Receive a signed Retell event
      responses:
        "202": { description: Event accepted or already processed }
        "401": { description: Invalid signature }
  /webhooks/twilio/status:
    post:
      operationId: receiveTwilioStatus
      tags: [calls]
      summary: Receive a Twilio call status callback
      responses:
        "202": { description: Event accepted or already processed }
        "401": { description: Invalid signature }
  /webhooks/calcom:
    post:
      operationId: receiveCalcomEvent
      tags: [appointments]
      summary: Receive a signed Cal.com booking event
      responses:
        "202": { description: Event accepted or already processed }
        "401": { description: Invalid signature }
...